requirements.txt
anthropic>=0.45.0
yfinance>=0.2.40
pandas>=2.2.0
pyyaml>=6.0
aiohttp>=3.9.0
